refactor: bundle component with style
