+++ client/src/components/NavOutlet.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Nav from './Nav';
+import Nav from './Nav/Nav';
 
 const NavOutlet = () => {
   return (
